move study password validation to function
